// src/app/domain/resources/repository.file.response.ts
export interface RepositoryFileResponse {

  uri: string;
  description: string;
  created: string;
  modified: string;
  name: string;
  owner: any;
  id: string;
  accessPermission: number;
  type: string;
  leaf: boolean;
  text: string;
  qtip: string;
  permission: number;
  parents: string;

}