// src/app/domain/resources/documents-repository-file.ts
export interface DocumentsRepositoryFile {
  uri: string;
  description: string;
  created: string;
  modified: string;
  name: string;
  owner: any;
  id: string;
  accessPermission: number;
  type: string;
  storageType: string;
  selected: boolean;
  mime: string;
  size: string;
  thumbnail: string;
  fileDownload: string;
  isLeaf: boolean;
  parents: any;
}